Move cutoff functions to scripts
